requirements.txt
Flask==3.1.0
pillow==11.1.0
numpy==2.1.3
opencv-python==4.11.0.86
tensorflow==2.19.0